requirements.txt
pyasn1==0.6.1
pyserial==3.5
PySimpleGUI==5.0.8
rsa==4.9
